feat(dashboard): PnL state & layout polish (saved from local Test)
